ApplicationInsight/python/requirements.txt
azure-monitor-opentelemetry
python-dotenv